+++ packages/cardano-services/src/Rewards/DbSyncRewardProvider/index.ts
@@ -1,4 +1,4 @@
-export * from './DbSyncRewards';
+export * from './DbSyncRewardsProvider';
 export * from './mappers';
 export * from './types';
 export * from './RewardsBuilder';
