+++ packages/cardano-services/src/StakePool/DbSyncStakePoolProvider/index.ts
@@ -1,4 +1,4 @@
-export * from './DbSyncStakePool';
+export * from './DbSyncStakePoolProvider';
 export * from './mappers';
 export * from './types';
 export * from './StakePoolBuilder';
